+++ react-native.js
@@ -8,6 +8,16 @@ module.exports = {
     // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-filename-extension.md
     // Cf. https://github.com/facebook/react-native/pull/5233
     'react/jsx-filename-extension': [2, { extensions: ['.js'] }],
-
+  },
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: [
+          '.js',
+          '.ios.js',
+          '.android.js',
+        ],
+      },
+    },
   },
 };